+++ src/coreclr/vm/callconvbuilder.hpp
@@ -78,7 +78,7 @@ namespace CallConv
         _In_ PCCOR_SIGNATURE pSig,
         _In_ ULONG cSig,
         _Inout_ CallConvBuilder *builder,
-        _Out_ UINT *errorResID);
+        _Out_ WORD *errorResID);
 
     //-------------------------------------------------------------------------
     // Gets the calling convention from the UnmanagedCallConv attribute
@@ -92,7 +92,7 @@ namespace CallConv
     HRESULT TryGetCallingConventionFromUnmanagedCallConv(
         _In_ MethodDesc* pMD,
         _Inout_ CallConvBuilder* builder,
-        _Out_opt_ UINT* errorResID);
+        _Out_opt_ WORD* errorResID);
 
     //-------------------------------------------------------------------------
     // Gets the unmanaged calling convention from the UnmanagedCallersOnly attribute.
